% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{count_space}
\alias{count_space}
\title{Count the spaces between two given strings}
\usage{
count_space(mychar, sep)
}
\arguments{
\item{mychar}{The character to check.}

\item{sep}{character for separation.}
}
\value{
character as title with '#' inserted.
}
\description{
Count the spaces between two given strings
}
