% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{dir2mm}
\alias{dir2mm}
\title{Display hierarchical structure of a directory in FreeMind mind map}
\usage{
dir2mm(
  from = ".",
  root = NA,
  dir_files = TRUE,
  dir_all = TRUE,
  dir_excluded = NA,
  md_maxlevel = ""
)
}
\arguments{
\item{from}{Character. The source text of the (R) Markdown syntax text, the R code, the FreeMind mind map code, or the path to the directory.}

\item{root}{Character. The string displayed as the root (center) of the mind map.}

\item{dir_files}{Logical. Whether to include files. If \code{FALSE}, only folders are included. If \code{TRUE}, folders and files are included.}

\item{dir_all}{Logical. Whether to include all files in a directory. If \code{FALSE}, only the names of visible files are included (following Unix-style visibility, that is files whose name does not start with a dot). If \code{TRUE}, all file names will be included.}

\item{dir_excluded}{Character. The directories which are not included in the output.}

\item{md_maxlevel}{Integer or ''. The maximum level of the markdown headings that are displayed in the mind map.}
}
\value{
FreeMind mind map code.
}
\description{
Display hierarchical structure of a directory in FreeMind mind map
}
\examples{
input <- system.file(package = "mindr")
dir2mm(input)
dir2mm(input, dir_files = FALSE, dir_all = TRUE, dir_excluded = "Meta")

output_txt <- dir2mm(input)
output <- tempfile(pattern = "file", tmpdir = tempdir(), fileext = ".mm")
writeLines(output_txt, output, useBytes = TRUE)
message("Input:  ", input, "\nOutput: ", output)
# file.show(output) # Open the output file file.remove(output) # remove the
# output file
}
