% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{dir2r}
\alias{dir2r}
\title{Convert a hierarchical directory into R code}
\usage{
dir2r(
  from = ".",
  dir_files = TRUE,
  dir_all = TRUE,
  dir_excluded = NA,
  r_seclabel = " --------",
  r_chunkheading = FALSE
)
}
\arguments{
\item{from}{Character. The source text of the (R) Markdown syntax text, the R code, the FreeMind mind map code, or the path to the directory.}

\item{dir_files}{Logical. Whether to include files. If \code{FALSE}, only folders are included. If \code{TRUE}, folders and files are included.}

\item{dir_all}{Logical. Whether to include all files in a directory. If \code{FALSE}, only the names of visible files are included (following Unix-style visibility, that is files whose name does not start with a dot). If \code{TRUE}, all file names will be included.}

\item{dir_excluded}{Character. The directories which are not included in the output.}

\item{r_seclabel}{Character. The ending characters indicating sections in R Markdown.}

\item{r_chunkheading}{Logical. Whether process the chunk label as headings.}
}
\value{
Character, R code
}
\description{
Convert a hierarchical directory into R code
}
\examples{
input <- system.file(package = "mindr")
dir2r(input)
dir2r(input, dir_files = FALSE, dir_all = TRUE, dir_excluded = "Meta")

output_txt <- dir2r(input)
output <- tempfile(pattern = "file", tmpdir = tempdir(), fileext = ".R")
writeLines(output_txt, output, useBytes = TRUE)
message("Input:  ", input, "\nOutput: ", output)
# system2('open', input) # Open the input dir in explorer.  file.show(output) #
# Open the output file file.remove(output) # remove the output file
}
