% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{dir4}
\alias{dir4}
\title{Convert a folder structure into a mindmap.}
\usage{
dir4(path = getwd(), savefile = TRUE, savefilename = "mindr.mm", output = c("mm", 
    "txt", "md", "Rmd"), backup = TRUE, dir_files = FALSE)
}
\arguments{
\item{path}{character. the path of the folder.}

\item{savefile}{logical. Whether to save the output as a file.}

\item{savefilename}{character. Valid when savefile == TRUE.}

\item{output}{a file with the folder structure.}

\item{backup}{logical. Whether the existing target file, if any, should be saved as backup.}

\item{dir_files}{logical. Whether to display files besides folders.}
}
\value{
a mindmap file, which can be viewed by common mindmap software, such as 'FreeMind' (\url{http://freemind.sourceforge.net/wiki/index.php/Main_Page}) and 'XMind' (\url{http://www.xmind.net}).
}
\description{
Convert a folder structure into a mindmap.
}
