% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_body}
\alias{get_body}
\title{get the body out of given strings}
\usage{
get_body(pattern = "^#[^ ]*", text)
}
\arguments{
\item{pattern}{The definition of the body text}

\item{text}{the given strings}
}
\value{
integer. the index of the body text in the given strings.
}
\description{
get the body out of given strings
}
