% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_eqloc}
\alias{get_eqloc}
\title{Get the index of equations in a string vector}
\usage{
get_eqloc(eq_begin, eq_end)
}
\arguments{
\item{eq_begin}{the beginning index of an equation}

\item{eq_end}{the end index of an equation}
}
\value{
a index vector
}
\description{
Get the index of equations in a string vector
}
