% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_filename_ext}
\alias{get_filename_ext}
\title{get the file name extention}
\usage{
get_filename_ext(filename)
}
\arguments{
\item{filename}{character, the file name}
}
\value{
character, the file name extention
}
\description{
get the file name extention
}
