% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_foldername}
\alias{get_foldername}
\title{Get the folder name of a given complete path}
\usage{
get_foldername(path)
}
\arguments{
\item{path}{The complete path}
}
\value{
The folder name
}
\description{
Get the folder name of a given complete path
}
\examples{
get_foldername("c:/dropbox/temp")
}
