% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_heading}
\alias{get_heading}
\title{get the headings out of given strings}
\usage{
get_heading(pattern = "^#+ ", text)
}
\arguments{
\item{pattern}{The definition of the headings}

\item{text}{the given strings}
}
\value{
integer. the index of the headings in the given strings.
}
\description{
get the headings out of given strings
}
