% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{guess_type}
\alias{guess_type}
\title{Guess the type of input or output}
\usage{
guess_type(from)
}
\arguments{
\item{from}{The source text}
}
\value{
the type, including 'dir', 'mindmap', 'R', 'markdown'.
}
\description{
Guess the type of input or output
}
