% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{list2heading}
\alias{list2heading}
\title{convert lists to headings in a text}
\usage{
list2heading(text)
}
\arguments{
\item{text}{the given strings}
}
\value{
integer. the index of the headings in the given strings.
}
\description{
convert lists to headings in a text
}
