% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{markmap}
\alias{markmap}
\title{Create a markmap widget}
\usage{
markmap(folder = "mm", remove_curly_bracket = FALSE, width = NULL, height = NULL, 
    elementId = NULL, options = markmapOption())
}
\arguments{
\item{width}{the width of the markmap}

\item{height}{the height of the markmap}

\item{options}{the markmap options}

\item{path}{the path of markdown file}
}
\value{
A HTML widget object rendered from a given document.
}
\description{
This function creates a markmap widget using htmlwidgets.
The widget can be rendered on HTML pages generated from
R Markdown,Shiny,or other applications.
}
\examples{
folder <- system.file("examples/md", package = "mindr")
markmap(folder = folder)
markmap(folder = folder, remove_curly_bracket = TRUE)
}
