% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{markmapOutput}
\alias{markmapOutput}
\title{Shiny bindings for markmap}
\usage{
markmapOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{output variable to read from}

\item{width}{Must be a valid CSS unit (like \code{'100\%'}, \code{'400px'}, \code{'auto'}) or a number, which will be coerced to a string and have \code{'px'} appended.}

\item{height}{See 'width'.}
}
\description{
Output function for using markmap within Shiny applications and interactive Rmd documents. This function is taken from \url{https://github.com/seifer08ms/Rmarkmap}.
}
