% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindr.R
\name{mdtxt2mmtxt}
\alias{mdtxt2mmtxt}
\title{Convert markdown text to mindmap text.}
\usage{
mdtxt2mmtxt(title = "my title", mdtxt = "", keep_eq = FALSE)
}
\arguments{
\item{title}{character. The title of the output file.}

\item{mdtxt}{character. The markdown text to convert.}

\item{keep_eq}{logical. whether to keep LaTeX equations.}
}
\value{
a mindmap text.
}
\description{
Convert markdown text to mindmap text.
}
\examples{
mdtxt2mmtxt(mdtxt = c("# Chapter 1", "## Section 1.1", "## Section 1.2"))
}
