% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{mindr}
\alias{mindr}
\title{mindr()}
\usage{
mindr(title = "my title", folder = "mm", headerfile = FALSE, 
    task = "markdown to mindmap")
}
\arguments{
\item{title}{character. The title of the output file.}

\item{folder}{character. The folder which contains the input file(s).}

\item{headerfile}{logical. Whether to save the headers into a markdown file.}

\item{task}{character. Its value can be:
- 'markdown to mindmap' (default)
- 'mindmap to markdown'}
}
\value{
a mindmap file,which can be viewed by mindmap software, such as FreeMind or Xmind.
}
\description{
mindr()
}
\examples{
mindr()
}
