% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R
\name{mmm}
\alias{mmm}
\title{Convert almost any file into mind map.}
\usage{
mmm(
  input_file = NA,
  output_type = c("widget", "mindmap", "markdown"),
  root = NA,
  md_list = FALSE,
  md_eq = FALSE,
  md_braces = FALSE,
  md_bookdown = FALSE,
  md_maxlevel = "",
  r_seclabel = " --------",
  r_chunkheading = FALSE,
  widget_name = NA,
  widget_width = NULL,
  widget_height = NULL,
  widget_elementId = NULL,
  widget_options = markmapOption(preset = "colorful")
)
}
\arguments{
\item{input_file}{Character. The path to the file for input.}

\item{output_type}{Character. The type of the output. It can be \code{'widget', 'mindmap', 'markdown', 'R', 'dir'}. The default value is \code{'widget'}.}

\item{root}{Character. The string displayed as the root (center) of the mind map.}

\item{md_list}{Logical. whether to process lists like headings in the Markdown input.}

\item{md_eq}{Logical. Whether to include LaTeX equations in the Markdown input when converted to other formats.}

\item{md_braces}{Logical. Whether to remove \href{https://bookdown.org/yihui/bookdown/cross-references.html}{{#ID}} in the headings of the markdown file (usually in a \href{https://github.com/rstudio/bookdown}{bookdown}> project.}

\item{md_bookdown}{Logical. Whether the R Markdown syntax text is in bookdown style, i.e. \code{# (PART), # (APPENDIX)}, and \code{# References} as an upper level of the Level 1 heading.}

\item{md_maxlevel}{Integer or ''. The maximum level of the markdown headings that are displayed in the mind map.}

\item{r_seclabel}{Character. The ending characters indicating sections in R Markdown.}

\item{r_chunkheading}{Logical. Whether process the chunk label as headings.}

\item{widget_name}{Character. The name of the html widget.}

\item{widget_width}{Numeric. The width of the widget.}

\item{widget_height}{Numeric. The height of the widget.}

\item{widget_elementId}{Character. The ID of teh Widget.}

\item{widget_options}{List. Options for the markmap widget. It should be a list passed from the \code{markmapOption()} function.}
}
\value{
Desired output.
}
\description{
Convert almost any file into mind map.
}
\details{
The input file type could be .md, .Rmd, .R, .mm, .pdf, .docx, .html, .rtf, .odt, .epub, .tex, and any other types which pandoc can convert from. See \href{https://pandoc.org/}{pandoc} for more details.
}
\examples{
# pdf: here we use the pdf files in the R installation
input <- file.path(R.home(), "doc/manual/R-intro.pdf")
mmm(input)
input <- file.path(R.home(), "doc/manual/R-data.pdf")
mmm(input, root = "R-data", md_maxlevel = 2)

# MS Word: here we use a .docx document shipped by the 'officer' package input
# <- system.file('doc_examples/example.docx', package = 'officer') mmm(input)

# HTML: here we use the R-FQA webpage
input <- file.path(R.home(), "doc/html/rw-FAQ.html")
mmm(input)
}
