% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{outline}
\alias{outline}
\title{Outline
Extract headers of markdown or rmarkdown files as an outline}
\usage{
outline(folder = "mm", remove_curly_bracket = FALSE, inputformat = "markdown", 
    savefile = FALSE, savefilename = "outline.md")
}
\arguments{
\item{folder}{character. The folder which contains the input file(s).}

\item{remove_curly_bracket}{logical.}

\item{inputformat}{character. Its value can be:
- 'markdown' (default)
- 'mindmap'}

\item{savefile}{logical}

\item{savefilename}{character. Valid when savefile == TRUE}
}
\value{
outline of a markdown document or book.
}
\description{
Outline
Extract headers of markdown or rmarkdown files as an outline
}
\examples{
outline()
}
