% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R
\name{outline_pdf}
\alias{outline_pdf}
\title{Extract the outline from pdf toc, and output as Markdown}
\usage{
outline_pdf(input_toc)
}
\arguments{
\item{input_toc}{Character. The table of contents (TOC) of a pdf file, extracted by \code{pdftools::pdf_toc()}.}
}
\value{
Character, showing the TOC in Markdown.
}
\description{
Extract the outline from pdf toc, and output as Markdown
}
\examples{
input <- file.path(R.home(), "doc/manual/R-intro.pdf")
input_toc <- pdftools::pdf_toc(input)
outline_pdf(input_toc)
}
