% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{r2dir}
\alias{r2dir}
\title{Generate hierarchical directories according to the outline of R code}
\usage{
r2dir(
  from = NA,
  dir_to = NA,
  md_list = FALSE,
  md_bookdown = TRUE,
  dir_quiet = FALSE
)
}
\arguments{
\item{from}{Character. The source text of the (R) Markdown syntax text, the R code, the FreeMind mind map code, or the path to the directory.}

\item{dir_to}{Character. The path of the output directory.}

\item{md_list}{Logical. whether to process lists like headings in the Markdown input.}

\item{md_bookdown}{Logical. Whether the R Markdown syntax text is in bookdown style, i.e. \code{# (PART), # (APPENDIX)}, and \code{# References} as an upper level of the Level 1 heading.}

\item{dir_quiet}{Logical. Whether to display the results of generated directories.}
}
\value{
Directory generated.
}
\description{
Generate hierarchical directories according to the outline of R code
}
\examples{
input <- system.file("examples/mindr-r.R", package = "mindr")
input_txt <- readLines(input, encoding = "UTF-8")
output <- file.path(tempdir(), "mindr")
r2dir(input_txt, output)
message("Input:  ", input, "\nOutput: ", output)
# file.show(input) # Open the input file system2('open', output) # Open the
# output dir in explorer.  unlink(output, recursive = TRUE) # remove the output
# file
}
