% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindr.R
\name{r2md}
\alias{r2md}
\title{Convert .R scripts into a .md/.Rmd file}
\usage{
r2md(filepattern = "*.R$", path = ".", savefilename = NA, backup = TRUE, body = "#' ")
}
\arguments{
\item{filepattern}{the pattern of the script file names}

\item{path}{the path of the folder which contains the .R scripts}

\item{savefilename}{the destinated file name}

\item{backup}{logical. whether backup the existent file}

\item{body}{the indicator of the body text}
}
\value{
a markdown file
}
\description{
Convert .R scripts into a .md/.Rmd file
}
\examples{
r2md()
}
