% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindr.R
\name{r2rmd}
\alias{r2rmd}
\title{Convert .R scripts into a .Rmd file}
\usage{
r2rmd(filepattern = "*.R$", path = ".", savefilename = NA)
}
\arguments{
\item{filepattern}{the pattern of the script file names}

\item{path}{the path of the folder which contains the .R scripts}

\item{savefilename}{the destinated file name}
}
\value{
an R markdown file
}
\description{
Convert .R scripts into a .Rmd file
}
\examples{
path <- system.file("examples/r", package = "mindr")
r2rmd(path = path)
}
