% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{rename2}
\alias{rename2}
\title{Rename a file automatically with a time stamp}
\usage{
rename2(filename, connect = "-")
}
\arguments{
\item{filename}{character.}

\item{connect}{the connecting character in the time stamp}
}
\value{
a new file name
}
\description{
Rename a file automatically with a time stamp
}
