% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindr.R
\name{renderMarkmap}
\alias{renderMarkmap}
\title{Shiny bindings for markmap}
\usage{
renderMarkmap(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that generates a markmap}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This is useful if you want to save an expression in a variable.}
}
\description{
Render function for using markmap within Shiny applications and interactive Rmd documents. This function is taken from \url{https://github.com/seifer08ms/Rmarkmap}.
}
