% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindr.R
\name{rmd2r}
\alias{rmd2r}
\title{Convert .md or .Rmd files into a .R script}
\usage{
rmd2r(filepattern = "*.[R]*md$", path = ".", savefile = TRUE, savefilename = NA, 
    backup = TRUE, heading = " --------", chunkheading = FALSE)
}
\arguments{
\item{filepattern}{the pattern of the file names}

\item{path}{the path of the folder which contains the .Rmd or .md files}

\item{savefile}{logical. Whether to save the output as a file.}

\item{savefilename}{the destinated file name}

\item{backup}{logical. whether backup the existent file}

\item{heading}{the indicator of the headings}

\item{chunkheading}{logical. whether treat chunk options as headings (ending with ----)}
}
\value{
a .R script
}
\description{
Convert .md or .Rmd files into a .R script
}
\examples{
path <- system.file("examples/r", package = "mindr")
rmd2r(path = path)
}
