% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{rmvcode}
\alias{rmvcode}
\title{Check whether a digital number is within a given range}
\usage{
rmvcode(index, loc)
}
\arguments{
\item{index}{integer. a row number in a markdown file}

\item{loc}{integer vector. the row numbers of the code block indicator, e.g.  triple backsticks}
}
\value{
logical.
}
\description{
Check whether a digital number is within a given range
}
