% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{writeLines2}
\alias{writeLines2}
\title{Write txt files avoiding overwriting existent files.}
\usage{
writeLines2(text, filename, backup = TRUE)
}
\arguments{
\item{text}{The text to write.}

\item{filename}{The destinated file name}

\item{backup}{Logical.}
}
\value{
a txt file
}
\description{
Write txt files avoiding overwriting existent files.
}
