% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{dir2}
\alias{dir2}
\title{Convert a folder structure into a mindmap.}
\usage{
dir2(path = getwd(), savefilename = "mindr", output = c("mm", "txt", "md"), 
    backup = TRUE)
}
\arguments{
\item{path}{character. the path of the folder.}

\item{savefilename}{character. Valid when savefile == TRUE.}

\item{output}{a file with the folder structure.}

\item{backup}{logical. Whether the existing target file, if any, should be saved as backup.}
}
\value{
a mindmap file, which can be viewed by common mindmap software, such as 'FreeMind' (\url{http://freemind.sourceforge.net/wiki/index.php/Main_Page}) and 'XMind' (\url{http://www.xmind.net}).
}
\description{
Convert a folder structure into a mindmap.
}
\examples{
dir2(NA)
}
