% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{mdtxt2mmtxt}
\alias{mdtxt2mmtxt}
\title{Convert Markdown text to FreeMind mind map text.}
\usage{
mdtxt2mmtxt(from = "", root = "root", md_eq = FALSE)
}
\arguments{
\item{from}{Character. The source text of the (R) Markdown syntax text, the R code, the FreeMind mind map code, or the path to the directory.}

\item{root}{Character. The string displayed as the root (center) of the mind map.}

\item{md_eq}{Logical. Whether to include LaTeX equations in the Markdown input when converted to other formats.}
}
\value{
a mindmap text.
}
\description{
Convert Markdown text to FreeMind mind map text.
}
