% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{mdtxt2mmtxt}
\alias{mdtxt2mmtxt}
\title{Convert markdown text to mindmap text.}
\usage{
mdtxt2mmtxt(title = "my title", mmtxt = "")
}
\arguments{
\item{title}{character. The title of the output file.}

\item{mmtxt}{character. The markdown text to convert.}
}
\value{
a mindmap text.
}
\description{
Convert markdown text to mindmap text.
}
\examples{
mdtxt2mmtxt(mmtxt = c("# Chapter 1", "## Section 1.1", "## Section 1.2"))
}
