% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo.R
\name{mm2md}
\alias{mm2md}
\title{Convert a mind map (.mm) into a markdown file.}
\usage{
mm2md(folder = "mm", savefile = TRUE, savefilename = "mm.md")
}
\arguments{
\item{folder}{character. The folder which contains the input file(s).}

\item{savefile}{logical}

\item{savefilename}{character. Valid when savefile == TRUE}
}
\value{
outline of a markdown document or book.
}
\description{
Convert a mind map (.mm) into a markdown file.
}
\examples{
mm2md()
}
