% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindr.R
\name{mm2r}
\alias{mm2r}
\title{Convert .mm into a .R script}
\usage{
mm2r(filepattern = "*.mm$", path = ".", savefile = TRUE, savefilename = NA, 
    backup = TRUE, heading = " --------")
}
\arguments{
\item{filepattern}{the pattern of the file names}

\item{path}{the path of the folder which contains the .Rmd or .md files}

\item{savefile}{logical. Whether to save the output as a file.}

\item{savefilename}{the destinated file name}

\item{backup}{logical. whether backup the existent file}

\item{heading}{the indicator of the headings}
}
\value{
a .R script
}
\description{
Convert .mm into a .R script
}
\examples{
path <- system.file("examples/mm", package = "mindr")
mm2r(path = path)
}
