% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindr.R
\name{r2md}
\alias{r2md}
\title{Convert .R scripts into a .md/.Rmd file}
\usage{
r2md(path = ".", filepattern = "*.R$", savefilename = NA, backup = TRUE, 
    heading = " --------$", body = "^#[^[:blank:]#]+")
}
\arguments{
\item{path}{the path of the folder which contains the .R scripts}

\item{filepattern}{the pattern of the script file names}

\item{savefilename}{the destinated file name}

\item{backup}{logical. whether backup the existent file}

\item{heading}{the indicator of the headings}

\item{body}{the indicator of the body text}
}
\value{
a markdown file
}
\description{
Convert .R scripts into a .md/.Rmd file
}
\examples{
r2md()
}
